orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "educonnect-orchestrator"
version = "0.1.0"
description = "Test coverage and CI/CD improvement orchestrator for EduConnect Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "EduConnect Team"}
]
keywords = ["testing", "ci-cd", "orchestration", "coverage"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "asyncio-throttle>=1.0.2",
    "anyio>=4.2.0",
    "aiofiles>=23.2.1",
    "sh>=2.0.6",
    "psutil>=5.9.7",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "gitpython>=3.1.41",
    "pygithub>=2.1.1",
    "httpx>=0.26.0",
    "jinja2>=3.1.2",
    "structlog>=24.1.0",
    "tenacity>=8.2.3",
    "docker>=7.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.11",
    "black>=23.12.1",
    "isort>=5.13.2",
]
notifications = [
    "slack-sdk>=3.26.1",
    "sendgrid>=6.11.0",
]
kubernetes = [
    "kubernetes>=28.1.0",
]

[project.scripts]
orchestrate = "orchestrator.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
